refactor: extracting methods and subview from the root
