+++ ios-app/YouPlay/Core/Root/View/RootView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 
 struct RootView: View {
     @StateObject private var viewModel = RootViewModel(spotifyController: SpotifyController())
-    @State private var isShowingSongDetail = false
+    @State private var showPlaybackPlayer = false
     @State private var keyboardHeight: CGFloat = 0.0
 
     var body: some View {
@@ -34,79 +34,24 @@ struct RootView: View {
                                 }
                             }
 
-                        // Song player crumbar
-                        if let song = viewModel.song,
-                           viewModel.currentUser != nil,
-                           keyboardHeight == 0.0 // hide when keyboard is open
-                        {
-                            VStack {
-                                Spacer()
-
-                                Button(action: {
-                                    isShowingSongDetail.toggle()
-                                }) {
-                                    HStack {
-                                        // song info
-                                        HStack(spacing: 12) {
-                                            AlbumImageView(
-                                                image: song.album.images.first,
-                                                width: 42.0,
-                                                height: 42.0,
-                                                borderRadius: .small
-                                            )
-
-                                            VStack(alignment: .leading) {
-                                                Text(song.name)
-                                                    .font(.callout)
-                                                    .fontWeight(.semibold)
-                                                    .lineLimit(1)
-
-                                                Text(song.artists
-                                                    .compactMap { artist in artist.name }
-                                                    .joined(separator: ", ")
-                                                )
-                                                .font(.caption)
-                                                .foregroundStyle(.gray)
-                                                .lineLimit(1)
-                                            }
-                                        }
-
-                                        Spacer()
-
-                                        // play/pause button
-                                        Button {
-                                            viewModel.spotifyController.pauseOrPlay()
-                                        } label: {
-                                            Image(systemName: viewModel.isPaused ? "play.circle.fill" : "pause.circle.fill")
-                                                .foregroundColor(.primary)
-                                                .font(.system(size: 30))
-                                        }
-                                        .padding(.trailing, 8)
-                                    }
-                                    .padding(8)
-                                    .background(Color(.systemGray6))
-                                    .clipShape(RoundedRectangle(cornerRadius: 8))
-                                    .padding(.horizontal, 6)
-                                }
-                                .tint(.white)
-
-                                // the following spacers account for the TabBar
-                                Spacer()
-                                    .frame(width: UIScreen.main.bounds.width, height: 10)
-                                    .background(Color.black.opacity(0.8))
-
-                                Spacer()
-                                    .frame(width: UIScreen.main.bounds.width, height: 45)
-                            }
-                        }
+                        CrumbBarPlayerView(
+                            viewModel: viewModel,
+                            showPlaybackPlayer: $showPlaybackPlayer,
+                            keyboardHeight: $keyboardHeight
+                        )
                     }
                 }
                 .tint(.green)
             }
         }
-        .sheet(isPresented: $isShowingSongDetail) {
+        .sheet(isPresented: $showPlaybackPlayer) {
             NavigationStack {
-                SongDetailView(viewModel: viewModel)
+                PlaybackPlayerView(viewModel: viewModel)
+                    .toolbar {
+                        ToolbarItem(placement: .topBarTrailing) {
+                            PlaybackPlayerMenuView(viewModel: viewModel)
+                        }
+                    }
             }
             .tint(.green)
         }
